pyproject.toml
[tool.poetry]
name = "vector-db-with-llm"
version = "0.1.0"
description = ""
authors = ["skip"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.7"
langchainhub = "^0.1.21"
tiktoken = "^0.8.0"
chromadb = "^0.5.18"
langchain-community = "^0.3.7"
bs4 = "^0.0.2"
python-dotenv = "^1.0.1"
pypdf = "^5.1.0"
sentence-transformers = "^3.3.1"
requests = "^2.27.1"
faiss-cpu = "^1.9.0.post1"
chardet = "^5.2.0"
streamlit = "^1.40.2"
streamlit-chat = "^0.1.1"
docx2txt = "^0.8"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
gunicorn = "^23.0.0"
pytz = "^2024.2"
httpx = "^0.27.2"
pytest = "^8.3.3"
pytest-cov = "^6.0.0"
pyyaml = "^6.0.2"
jsondiff = "^2.2.1"
elasticsearch = "7.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
